+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Upcoming
+# 4.10.1
 ## New Feature
 - Add custom cell option to data table
 - Add toolbar option in data table